Move firebaseConfig into firebase directory
