+++ src/services/firebase/firebase.js
@@ -2,7 +2,7 @@ import app from 'firebase/app';
 import 'firebase/auth';
 import 'firebase/firestore';
 
-import firebaseConfig from '../../config/firebaseConfig';
+import firebaseConfig from './config';
 
 class Firebase {
   constructor() {
